+++ R/wq.R
@@ -12,7 +12,7 @@
 #'   "SKT" (Spring Kodiak Trawl, \code{\link{SKT}}),
 #'   "SLS" (Smelt Larva Survey, \code{\link{SLS}}),
 #'   "Baystudy" (Bay Study, \code{\link{baystudy}}),
-#'   "USGS" (USGS San Francisco Bay Surveys, \code{\link{USGS}}),
+#'   "USGS_SFBS" (USGS San Francisco Bay Surveys, \code{\link{USGS_SFBS}}),
 #'   "USBR" (United States Bureau of Reclamation Sacramento Deepwater Ship Channel data, \code{\link{USBR}}),
 #'   "Suisun" (Suisun Marsh Fish Study, \code{\link{suisun}})
 #'   "YBFMP" (Yolo Bypass Fish Monitoring Program, \code{\link{YBFMP}}), and
@@ -25,7 +25,7 @@
 #' @examples
 #' Data <- wq(Sources = c("EMP", "STN", "FMWT", "EDSM",
 #' "DJFMP", "SDO", "SKT", "SLS",
-#' "20mm", "Suisun", "Baystudy", "USBR", "USGS", "YBFMP", "USGS_CAWSC"))
+#' "20mm", "Suisun", "Baystudy", "USBR", "USGS_SFBS", "YBFMP", "USGS_CAWSC"))
 #' @export
 
 wq <- function(Sources=NULL,
@@ -35,10 +35,14 @@ wq <- function(Sources=NULL,
 
 # Check arguments ---------------------------------------------------------
 
+  if("USGS"%in%(Sources)){
+    stop('The "USGS" data source has been renamed to "USGS_SFBS" because of the inclusion of an additional USGS dataset, "USGS_CAWSC".')
+  }
+
 if(is.null(Sources) | !all(Sources%in%c("EMP", "STN", "FMWT", "EDSM", "DJFMP", "SDO", "SKT", "SLS",
-                                        "20mm", "Suisun", "Baystudy", "USBR", "USGS", "YBFMP", "USGS_CAWSC"))){
+                                        "20mm", "Suisun", "Baystudy", "USBR", "USGS_SFBS", "YBFMP", "USGS_CAWSC"))){
   stop('You must specify the data sources you wish to include. Choices include
-  c("EMP", "STN", "FMWT", "EDSM", "DJFMP", "SDO", "SKT", "SLS", "20mm", "Suisun", "Baystudy", "USBR", "USGS", "YBFMP", "USGS_CAWSC")')
+  c("EMP", "STN", "FMWT", "EDSM", "DJFMP", "SDO", "SKT", "SLS", "20mm", "Suisun", "Baystudy", "USBR", "USGS_SFBS", "YBFMP", "USGS_CAWSC")')
 }
 
   # Set end year to current year if blank
@@ -102,8 +106,8 @@ if(is.null(Sources) | !all(Sources%in%c("EMP", "STN", "FMWT", "EDSM", "DJFMP", "
     WQ_list[["USBR"]]<-discretewq::USBR
   }
 
-  if("USGS"%in%Sources){
-    WQ_list[["USGS"]]<-discretewq::USGS
+  if("USGS_SFBS"%in%Sources){
+    WQ_list[["USGS_SFBS"]]<-discretewq::USGS_SFBS
   }
 
   if("YBFMP"%in%Sources){